% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{openmi.om.logger-class}
\alias{openmi.om.logger-class}
\alias{openmi.om.logger}
\title{The base class for logging component data.}
\value{
reference class of type openmi.om.linkableComponent
}
\description{
The base class for logging component data.
}

