% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{openmi.om.matrix-class}
\alias{openmi.om.matrix-class}
\alias{openmi.om.matrix}
\title{The base class for matrix/lokup table meta-model components.}
\value{
reference class of type openmi.om.matrix
}
\description{
The base class for matrix/lokup table meta-model components.
}

