% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{openmi.om.timeSeriesInput-class}
\alias{openmi.om.timeSeriesInput-class}
\alias{openmi.om.timeSeriesInput}
\title{The base class for timeseries meta-model components.}
\value{
reference class of type openmi.om.timeSeriesInput
}
\description{
The base class for timeseries meta-model components.
}

